Replace & deprecate event emitter interface
This adds a new field called `ondevicechange` to the media devices
interface. It is very similar to the native behavior, except it passes
useful things instead of an event. These are the same parameters you're
used to from `on('devicechange')` but in object form.

The reasoning for the change is twofold:

- Better TypeScript type validation
- Avoidance of `events` library

The second point is important. While the bundle cost of `events` is
relatively good (~2kb), it's an implicit dependency which bundlers have
started to gripe about. It makes packaging these libraries more
difficult, because you don't know what to list as a dependency or what
bundler will be consuming them.

In hindsight, I believe the event emitter was a poor design choice, and
I believe the best path forward is to drop it in favor of the mutable
field.