Replace dts-cli with manual toolchain
I've been using `vite` as a library bundler recently. It is very good
and solves 80% of the pain that drove me to `dts-cli`. The rest is Jest,
ESLint, prettier, lint-staged, husky, and so on. But those tools are
easy to manage.

This commit removes `dts-cli` and uses the underlying libraries for most
components. The end user shouldn't notice a difference.